# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RUFORUM MEL API",
        "description": "RUFORUM MEL API Documentation - V1.0.0",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/grad-track-api/api",
            "description": "Local server"
        },
        {
            "url": "https://skills-api.comfarnet.org/api",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/api/dynamic-save": {
            "post": {
                "tags": [
                    "Dynamic CRUD"
                ],
                "summary": "Create or update a record dynamically",
                "description": "If 'id' is provided, updates the existing record; otherwise creates a new one. Other fields map to model columns.",
                "operationId": "40764374974da2fc3ed6ea0c8fb2f3d2",
                "parameters": [
                    {
                        "name": "model",
                        "in": "query",
                        "description": "Eloquent model name to operate on, e.g. 'User'",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Primary key of record to update (omit to create new)",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "JSON body with key-value pairs for model fields",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Record created or updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request: missing parameters or save error"
                    },
                    "401": {
                        "description": "Unauthorized: authentication required"
                    }
                }
            }
        },
        "/api/dynamic-list": {
            "get": {
                "tags": [
                    "Dynamic CRUD"
                ],
                "summary": "List or search records dynamically",
                "description": "Retrieve a paginated list of records for the given model, with optional filters, sorting, and pagination.",
                "operationId": "e3f75fcd741eec947cd2a513f7a3cc57",
                "parameters": [
                    {
                        "name": "model",
                        "in": "query",
                        "description": "Eloquent model name (without namespace) to query, e.g. 'User'",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "<field>_like",
                        "in": "query",
                        "description": "Filter where <field> LIKE %value%",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "<field>_gt",
                        "in": "query",
                        "description": "Filter where <field> > value",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Column name to sort by",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Sort direction: 'asc' or 'desc'",
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request: missing or invalid parameters"
                    },
                    "401": {
                        "description": "Unauthorized: authentication required"
                    }
                }
            }
        },
        "/api/dynamic-delete": {
            "post": {
                "tags": [
                    "Dynamic CRUD"
                ],
                "summary": "Delete a record dynamically",
                "description": "Deletes the record with the given 'id' from the specified model.",
                "operationId": "d4f739a10ab046e46f530def03daaf2f",
                "parameters": [
                    {
                        "name": "model",
                        "in": "query",
                        "description": "Eloquent model name to delete from, e.g. 'User'",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Primary key of the record to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request: missing 'id' parameter"
                    },
                    "401": {
                        "description": "Unauthorized: authentication required"
                    },
                    "404": {
                        "description": "Not found: record does not exist"
                    }
                }
            }
        }
    },
...
}